services/python-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Requests==2.31.0
pydantic==2.5.0
sqlalchemy==2.0.23
gunicorn==21.2.0
PyJWT==2.10.1
cryptography==41.0.7
